Reorganize codebase: flatten module tree + add Federation/Ingestion/Phase6 systems
- Move llm/, hid_driver/, mcp_service/, raft_consensus/, agentic_players/,
  wasm_ebus_bridge/ from core/hypervisor/ into core/hypervisor/src/
- Move Dockerfile, docker-compose, scripts into scripts/
- Move CRATE_QUICK_REFERENCE.csv, gen_icon.py, init-db.sql into docs/ or scripts/
- Delete orphaned config.toml, diff.txt, nats.zip, GENETICS_SETUP_COMPLETE.txt
- Add federation/ (multi-agent p2p, hive, DNA, specialists, enterprise, graph)
- Add native_ingestion/ (shmem_capture, simd_xor_delta, SVD, Chebyshev, etc.)
- Add shaders/ (WGSL GPU kernels)
- Add extensions/win32_intercept/ (Python HID bridge)
- Add scripts/ (GGUF harvester, helix compiler, spatial kinetic orchestrator)
- Phase 6: 38 computational logic systems (768 tests, 0 fail)

+++ core/hypervisor/src/dopamine_system.rs
@@ -1,5 +1,4 @@
-use nervous_system::shared_memory::SynapseState;
-use anyhow::Result;
+use crate::autonomic_loop::SynapseState;
 
 pub struct DopamineSystem;
 
@@ -22,7 +21,7 @@ impl DopamineSystem {
             }
             DopamineEvent::ExecutionFailure(severity) => {
                 // Penalty for failures (Integrity drop)
-                let penalty = severity * 10;
+                let penalty = (severity as u32) * 10;
                 state.integrity_score = state.integrity_score.saturating_sub(penalty);
                 state.understanding_score = state.understanding_score.saturating_sub(5);
                 // Failures increase curiosity (need to figure out why it failed)
@@ -46,3 +45,4 @@ pub enum DopamineEvent {
     ExecutionFailure(u8),    // contains severity (1-10)
     InternalCoherenceCheck(bool),
 }
+
